pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-bot"
version = "0.1.0"
description = "Binance Futures Testnet trading bot CLI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]
license = { text = "MIT" }
authors = [
    { name = "Trading Bot Developer" },
]
keywords = ["binance", "trading", "cli", "testnet"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
trading-bot = "cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
